+++ Script/LoadFilterCommand.swift
@@ -22,17 +22,11 @@ class LoadFilterCommand: ScriptCommand {
         self.callback = callback
         self.pattern = pattern
         self.loadFilterType = loadFilterType
-
-        if self.loadFilterType == .RequireToday {
-            let today = TodayCommand(callback: self.callback)
-            self.pattern = today.pattern
-        }
-
     }
 
     func validate() -> Bool {
         do {
-            if( self.loadFilterType != .Clear) {
+            if( self.loadFilterType == .Required || self.loadFilterType == .Excluded) {
                 try self.regex = NSRegularExpression(pattern: self.pattern, options: [])
             }
             return true
@@ -42,24 +36,50 @@ class LoadFilterCommand: ScriptCommand {
         }
     }
 
+    func changesData() -> Bool {
+        false
+    }
+
+
     func run(logLines: inout [LogLine], runState : inout RunState) -> Bool {
         switch( self.loadFilterType ) {
 
         case .Required:
-            fallthrough
-        case .RequireToday:
-            self.callback.scriptUpdate(text: "Added filter: Lines loaded must match: \(self.pattern)")
             runState.filterRequired.append(self.regex!)
+            self.callback.scriptUpdate(text: "Added filter: Lines loaded must match: \(self.pattern)")
+            return true
+        case .RequireToday:
+            // We have to set this up now, not at validation time, so it gets the most recent dateFormat value
+            return self.setupToday(runState: runState)
         case .Excluded:
             self.callback.scriptUpdate(text: "Added filter: Lines loaded must not match: \(self.pattern)")
             runState.filterExcluded.append(self.regex!)
+            return true
         case .Clear:
             self.callback.scriptUpdate(text: "Cleared filters on lines loaded")
             runState.filterRequired.removeAll()
             runState.filterExcluded.removeAll()
+            return true
         }
 
-        return true
+    }
+
+    func setupToday(runState : RunState) -> Bool {
+        let now = Date()
+          let formatter = DateFormatter()
+          formatter.dateFormat = runState.dateFormat
+          let pattern = formatter.string(from: now)
+          self.pattern = pattern
+
+          do {
+            try self.regex = NSRegularExpression(pattern: self.pattern, options: [])
+            runState.filterRequired.append(self.regex!)
+            self.callback.scriptUpdate(text: "Added filter: Lines loaded must match: \(self.pattern)")
+            return true
+          } catch {
+            self.callback.scriptUpdate(text: "invalid regular expression: \(self.pattern)")
+            return false
+          }
     }
 
 }